实现了一些窗口事件

+++ Biscuit/src/Biscuit/Event/MouseEvent.h
@@ -26,8 +26,8 @@ namespace Biscuit
 		EVENT_CLASS_CATEGORY(EventCategoryInput | EventCategoryMouse)
 
 	private:
-		float m_X;
-		float m_Y;
+		double m_X;
+		double m_Y;
 	};
 
 	/// <summary>
